pyproject.toml
[project]
name = "imgenhancewebservice"
version = "0.1.0"
requires-python = "==3.10.*" 
dependencies = [
    "absl-py==2.3.1",
    "addict==2.4.0",
    "albucore==0.0.24",
    "albumentations==2.0.8",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "astunparse==1.6.3",
    "backports-asyncio-runner==1.2.0",
    "basicsr==1.4.2",
    "beautifulsoup4==4.14.3",
    "black==25.12.0",
    "cachetools==6.2.4",
    "certifi==2026.1.4",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "colorama==0.4.6",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "exceptiongroup==1.3.1",
    "facexlib==0.3.0",
    "fastapi==0.128.0",
    "filelock==3.20.2",
    "filterpy==1.4.5",
    "flake8==7.3.0",
    "flatbuffers==25.12.19",
    "fonttools==4.61.1",
    "fsspec==2025.12.0",
    "future==1.0.0",
    "gast==0.7.0",
    "gdown==5.2.0",
    "gfpgan==1.3.8",
    "google-api-core==2.28.1",
    "google-api-python-client==2.187.0",
    "google-auth==2.41.1",
    "google-auth-httplib2==0.3.0",
    "google-auth-oauthlib==1.2.3",
    "google-pasta==0.2.0",
    "googleapis-common-protos==1.72.0",
    "grpcio==1.76.0",
    "h11==0.16.0",
    "h5py==3.15.1",
    "hf-xet==1.2.0",
    "httpcore==1.0.9",
    "httplib2==0.31.0",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "huggingface-hub==1.2.4",
    "idna==3.11",
    "imageio==2.37.2",
    "iniconfig==2.3.0",
    "jinja2==3.1.6",
    "joblib==1.5.3",
    "keras==3.12.0",
    "kiwisolver==1.4.9",
    "lazy-loader==0.4",
    "libclang==18.1.1",
    "librt==0.7.7",
    "llvmlite==0.46.0",
    "lmdb==1.7.5",
    "lz4==4.4.5",
    "markdown==3.10",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "matplotlib==3.10.8",
    "mccabe==0.7.0",
    "mdurl==0.1.2",
    "ml-dtypes==0.5.4",
    "mpmath==1.3.0",
    "mtcnn==1.0.0",
    "mypy==1.19.1",
    "mypy-extensions==1.1.0",
    "namex==0.1.0",
    "networkx==3.4.2",
    "numba==0.63.1",
    "numpy>=2.2.6",
    "oauthlib==3.3.1",
    "opencv-python==4.12.0.88",
    "opencv-python-headless==4.12.0.88",
    "opt-einsum==3.4.0",
    "optree==0.18.0",
    "packaging==25.0",
    "pathspec==1.0.2",
    "pillow==12.1.0",
    "platformdirs==4.5.1",
    "pluggy==1.6.0",
    "proto-plus==1.27.0",
    "protobuf==6.33.2",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycodestyle==2.14.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pyflakes==3.4.0",
    "pygments==2.19.2",
    "pyparsing==3.3.1",
    "pysocks==1.7.1",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.21",
    "pytokens==0.3.0",
    "pyyaml==6.0.3",
    "realesrgan==0.3.0",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "retina-face==0.0.17",
    "rich==14.2.0",
    "rsa==4.9.1",
    "scikit-image==0.25.2",
    "scipy==1.15.3",
    "shellingham==1.5.4",
    "simsimd==6.5.12",
    "six==1.17.0",
    "soupsieve==2.8.1",
    "starlette==0.50.0",
    "stringzilla==4.6.0",
    "super-image==0.2.0",
    "sympy==1.14.0",
    "tb-nightly==2.21.0a20251023",
    "tensorboard==2.20.0",
    "tensorboard-data-server==0.7.2",
    "tensorflow==2.20.0",
    "termcolor==3.3.0",
    "tifffile==2025.5.10",
    "tomli==2.3.0",
    "torch==2.0.0+cu118",
    "torchvision==0.15.0+cu118",
    "tqdm==4.67.1",
    "typer-slim==0.21.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "uritemplate==4.2.0",
    "urllib3==2.6.3",
    "uvicorn==0.40.0",
    "watchfiles==1.1.1",
    "websockets==15.0.1",
    "werkzeug==3.1.5",
    "wrapt==2.0.1",
    "yapf==0.43.0",
]


[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu118" }
torchvision = { index = "pytorch-cu118" }
torchaudio = { index = "pytorch-cu118" }
